Made Amph specialHexes be supply symbols.
Made new alterSpecialHex to all for specialHexRemoval, without being just set again.

FIxed a bug with the playAs .php file.

Fixed commonSync specialHex insertion to be less ugly, not not ugly, just less.

css tweaks.

oh and the horrible touchMove bug that was causing the map to jump on touch events outside of the map